+++ packages/react/src/components/AnimatedHeader/components/HeaderTitle/HeaderTitle.tsx
@@ -43,15 +43,11 @@ const HeaderTitle: React.FC<HeaderTitleProps> = ({
       : 'en';
   const isNameFirst = NAME_FIRST_LANGS.includes(currentLang.slice(0, 2));
 
-  const headingCollapsed = `${blockClass}-collapsed`;
-  const headingExpanded = `${blockClass}-expanded`;
-
   return (
     <Tooltip align="bottom" label={`${welcomeText}, ${userName}`}>
       <h1
-        className={`${blockClass} ${
-          headerExpanded ? headingExpanded : headingCollapsed
-        }`}
+        className={blockClass}
+        data-expanded={headerExpanded}
         aria-label={ariaLabels?.welcome ?? `${welcomeText}, ${userName}`}>
         {isNameFirst ? (
           <>