+++ packages/web-components/src/components/global-header/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.4.0...@carbon-labs/wc-global-header@0.5.0) (2025-10-06)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.4.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.3.0...@carbon-labs/wc-global-header@0.4.0) (2025-09-26)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.3.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.2.0...@carbon-labs/wc-global-header@0.3.0) (2025-09-24)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.2.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.1.0...@carbon-labs/wc-global-header@0.2.0) (2025-09-22)
+
+
+### Bug Fixes
+
+* **mdx-components:** update package entrypoint ([#811](https://github.com/carbon-design-system/carbon-labs/issues/811)) ([763d680](https://github.com/carbon-design-system/carbon-labs/commit/763d6803e47446cbd336dec9a5be75afde23f6e7))
+
+
+
+
+
 # 0.1.0 (2025-09-17)